+++ frontend/src/assets/icons/IconIncome.tsx
