+++ frontend/src/assets/icons/IconUser.tsx
