+++ frontend/src/components/admin/GeneralInfo.tsx
@@ -4,7 +4,7 @@ import GeneralInfoUsers from "./GeneralInfoUsers";
 
 export default function GeneralInfo () {
     return(
-        <div className="flex justify-center py-8 gap-6 ">
+        <div className="flex py-8 gap-6 ">
           <GeneralInfoIncome/>
           <GeneralInfoOrder/>
           <GeneralInfoUsers/>
