--- frontend/src/pages/income/page.tsx
@@ -1,14 +1,0 @@
-"use client"
-import { useRouter } from "next/navigation";
-
-const Income = () => {
-    const router = useRouter()
-
-    return(
-        <div>
-            
-        </div>
-    )
-}
-
-export default Income;
