--- frontend/src/pages/product/page.tsx
@@ -1,14 +1,0 @@
-"use client"
-import { useRouter } from "next/navigation";
-
-const Product = () => {
-    const router = useRouter()
-
-    return(
-        <div>
-            
-        </div>
-    )
-}
-
-export default Product;
