--- frontend/src/pages/settings/page.tsx
@@ -1,14 +1,0 @@
-"use client"
-import { useRouter } from "next/navigation";
-
-const Settings = () => {
-    const router = useRouter()
-
-    return(
-        <div>
-            
-        </div>
-    )
-}
-
-export default Settings;
